Add tests for CustomerInfoList, remove CustomerInfoTable
