+++ src/components/Office/ShipmentApprovalPreview.test.jsx
@@ -4,7 +4,7 @@ import { mount } from 'enzyme';
 import ShipmentApprovalPreview from './ShipmentApprovalPreview';
 import ShipmentContainer from './ShipmentContainer';
 import AllowancesList from './DefinitionLists/AllowancesList';
-import CustomerInfoTable from './CustomerInfoTable';
+import CustomerInfoList from './DefinitionLists/CustomerInfoList';
 
 import { SHIPMENT_OPTIONS } from 'shared/constants';
 
@@ -220,7 +220,7 @@ describe('Shipment preview modal', () => {
     expect(wrapper.find(ShipmentApprovalPreview).exists()).toBe(true);
     expect(wrapper.find(ShipmentContainer).exists()).toBe(true);
     expect(wrapper.find(AllowancesList).exists()).toBe(true);
-    expect(wrapper.find(CustomerInfoTable).exists()).toBe(true);
+    expect(wrapper.find(CustomerInfoList).exists()).toBe(true);
 
     expect(wrapper.find('h4').at(0).text()).toEqual('Household goods');
     expect(wrapper.find('h4').at(2).text()).toEqual('Non-temp storage release');
@@ -244,7 +244,7 @@ describe('Shipment preview modal', () => {
     expect(wrapper.find(ShipmentApprovalPreview).exists()).toBe(true);
     expect(wrapper.find(ShipmentContainer).exists()).toBe(true);
     expect(wrapper.find(AllowancesList).exists()).toBe(true);
-    expect(wrapper.find(CustomerInfoTable).exists()).toBe(true);
+    expect(wrapper.find(CustomerInfoList).exists()).toBe(true);
   });
 
   it('renders the buttons successfully', () => {
